added basic password reset

+++ server/controllers/user.controller.js
@@ -4,6 +4,9 @@ const _ = require('lodash');
 
 const User = mongoose.model('User');
 const fileHelper = require('../util/file');
+const crypto = require('crypto');
+const nodemailer = require('nodemailer');
+
 
 module.exports.register = (req, res, next) => {
     try {
@@ -136,9 +139,13 @@ module.exports.updateUserProfile = (req, res, next) => {
                     }
                     foundedObject.save((err, updatedObject) => {
                         if (err) {
-                            res.status(500).send({msg: err});
+                            res.status(500).send({
+                                msg: err
+                            });
                         } else {
-                            res.send({msg: 'Profile Updated Successfully!'})
+                            res.send({
+                                msg: 'Profile Updated Successfully!'
+                            })
                         }
                     })
                 }
@@ -151,7 +158,7 @@ module.exports.updateUserProfile = (req, res, next) => {
 
 module.exports.updateProfileImage = (req, res, next) => {
     try {
-       
+
         const id = req._id;
         User.findOne({
             _id: id
@@ -164,31 +171,36 @@ module.exports.updateProfileImage = (req, res, next) => {
                 } else {
                     let imagePath = req.body.image;
                     if (req.file) {
-                      const url = req.protocol + "://" + req.get("host");
-                      imagePath = url + "/images/" + req.file.filename;
-                    } 
-                    
+                        const url = req.protocol + "://" + req.get("host");
+                        imagePath = url + "/images/" + req.file.filename;
+                    }
+
                     if (!imagePath) {
                         if (foundedObject.imagePath) {
-                            fileHelper.deleteFile('images/'+foundedObject.imagePath.split('images/')[1]);
+                            fileHelper.deleteFile('images/' + foundedObject.imagePath.split('images/')[1]);
                             foundedObject.imagePath = "";
                         }
                         foundedObject.imagePath = "";
                     }
                     if (imagePath) {
-                        console.log("imagePath", imagePath.split('images/')[1])
-                        console.log("foundedObject imagePath", foundedObject.imagePath.split('images/')[1])
-                        if (foundedObject.imagePath && imagePath.split('images/')[1]!==foundedObject.imagePath.split('images/')[1]) {
-                            fileHelper.deleteFile('images/'+foundedObject.imagePath.split('images/')[1]);
+                        // console.log("imagePath", imagePath.split('images/')[1])
+                        // console.log("foundedObject imagePath", foundedObject.imagePath.split('images/')[1])
+                        if (foundedObject.imagePath && imagePath.split('images/')[1] !== foundedObject.imagePath.split('images/')[1]) {
+                            fileHelper.deleteFile('images/' + foundedObject.imagePath.split('images/')[1]);
                             foundedObject.imagePath = imagePath;
                         }
                         foundedObject.imagePath = imagePath;
                     }
                     foundedObject.save((err, updatedObject) => {
                         if (err) {
-                            res.status(500).send({msg: err});
+                            res.status(500).send({
+                                msg: err
+                            });
                         } else {
-                            res.send({msg: 'Image uploaded successfully!', imagePath: updatedObject['imagePath']})
+                            res.send({
+                                msg: 'Image uploaded successfully!',
+                                imagePath: updatedObject['imagePath']
+                            })
                         }
                     })
                 }
@@ -213,23 +225,28 @@ module.exports.removeProfileImage = (req, res, next) => {
                 } else {
                     let imagePath = req.body.image;
                     if (req.file) {
-                      const url = req.protocol + "://" + req.get("host");
-                      imagePath = url + "/images/" + req.file.filename;
-                    } 
-                    
+                        const url = req.protocol + "://" + req.get("host");
+                        imagePath = url + "/images/" + req.file.filename;
+                    }
+
                     if (!imagePath) {
                         if (foundedObject.imagePath) {
-                            fileHelper.deleteFile('images/'+foundedObject.imagePath.split('images/')[1]);
+                            fileHelper.deleteFile('images/' + foundedObject.imagePath.split('images/')[1]);
                             foundedObject.imagePath = "";
                         }
                         foundedObject.imagePath = "";
                     }
-                
+
                     foundedObject.save((err, updatedObject) => {
                         if (err) {
-                            res.status(500).send({msg: err});
+                            res.status(500).send({
+                                msg: err
+                            });
                         } else {
-                            res.send({msg: 'Image removed successfully!', imagePath: updatedObject['imagePath']})
+                            res.send({
+                                msg: 'Image removed successfully!',
+                                imagePath: updatedObject['imagePath']
+                            })
                         }
                     })
                 }
@@ -270,7 +287,8 @@ module.exports.changePassword = (req, res, next) => {
                         message: 'Paswords do no match'
                     });
                 }
-                user.password = req.body.newPassword;
+                // user.password = req.body.newPassword;
+                user.password = User.hashPassword(req.body.newPassword)
                 user.save((err, doc) => {
                     if (!err)
                         return res.status(200).send({
@@ -309,4 +327,146 @@ module.exports.applyLeave = async (req, res, next) => {
         console.log('Cannot find User', error);
         return next(error);
     }
+}
+
+module.exports.ResetPassword = async (req, res) => {
+    if (!req.body.email) {
+        return res.status(500).json({
+            message: 'Email is required'
+        });
+    }
+    const user = await User.findOne({
+        email: req.body.email
+    });
+    if (!user) {
+        return res.status(409).json({
+            message: 'Email does not exist'
+        });
+    }
+    // let resettoken = new passwordResetToken({
+    //     _userId: user._id,
+    //     resettoken: crypto.randomBytes(16).toString('hex')
+    // });
+    user.resettoken = crypto.randomBytes(16).toString('hex')
+    user.save(function (err) {
+        if (err) {
+            return res.status(500).send({
+                msg: err.message
+            });
+        }
+
+        User.find({
+            _id: user._id,
+            resettoken: {
+                $ne: user.resettoken
+            }
+        }).remove().exec();
+        res.status(200).json({
+            message: 'Reset Password successfully.'
+        });
+        let transporter = nodemailer.createTransport({
+            service: 'Gmail',
+            port: 465,
+            auth: {
+                user: 'manjeetdhimaan60@gmail.com',
+                pass: 'lpaqbtmffjmepylc'
+            }
+        });
+        let mailOptions = {
+            to: user.email,
+            from: user.email,
+            subject: 'Employee Management Password Reset',
+            text: 'You are receiving this because you (or someone else) have requested the reset of the password for your account.\n\n' +
+                'Please click on the following link, or paste this into your browser to complete the process:\n\n' +
+                'http://localhost:4500/response-reset-password/' + user.resettoken + '\n\n' +
+                'If you did not request this, please ignore this email and your password will remain unchanged.\n'
+        }
+        transporter.sendMail(mailOptions, (err, info) => {})
+    })
+}
+
+
+module.exports.ValidPasswordToken = async (req, res) => {
+    if (!req.body.resettoken) {
+        return res
+            .status(500)
+            .json({
+                message: 'Token is required'
+            });
+    }
+    const user = await passwordResetToken.findOne({
+        resettoken: req.body.resettoken
+    });
+    if (!user) {
+        return res
+            .status(409)
+            .json({
+                message: 'Invalid URL'
+            });
+    }
+    User.findOneAndUpdate({
+        _id: user._userId
+    }).then(() => {
+        res.status(200).json({
+            message: 'Token verified successfully.'
+        });
+    }).catch((err) => {
+        return res.status(500).send({
+            msg: err.message
+        });
+    });
+}
+
+module.exports.NewPassword = async (req, res) => {
+    passwordResetToken.findOne({
+        resettoken: req.body.resettoken
+    }, function (err, userToken, next) {
+        if (!userToken) {
+            return res
+                .status(409)
+                .json({
+                    message: 'Token has expired'
+                });
+        }
+
+        User.findOne({
+            _id: userToken._userId
+        }, function (err, userEmail, next) {
+            if (!userEmail) {
+                return res
+                    .status(409)
+                    .json({
+                        message: 'User does not exist'
+                    });
+            }
+            return bcrypt.hash(req.body.newPassword, 10, (err, hash) => {
+                if (err) {
+                    return res
+                        .status(400)
+                        .json({
+                            message: 'Error hashing password'
+                        });
+                }
+                userEmail.password = hash;
+                userEmail.save(function (err) {
+                    if (err) {
+                        return res
+                            .status(400)
+                            .json({
+                                message: 'Password can not reset.'
+                            });
+                    } else {
+                        userToken.remove();
+                        return res
+                            .status(201)
+                            .json({
+                                message: 'Password reset successfully'
+                            });
+                    }
+
+                });
+            });
+        });
+
+    })
 }
